Fix FETools::extrapolate for block vectors
